Create linear_algebra.py
This module contains the algorithms for dealing with vector-like objects. It is too naive to say that it is a (real) vector space algorithm, since we can never really construct a (real) vector space in computers, since algebraic closure will never be attained with finite set --floating point numbers are finite.

In this algorithm, we will define a vector as a list of numbers. In general the numbers are floating-point numbers. However, integers are also compatible.